INCOMPLETE: Start middleware layer
